app/requirements.txt
streamlit>=1.32.0
google-cloud-bigquery>=3.41.0
google-auth>=2.49.0
pandas>=2.0.0
plotly>=5.20.0
python-dotenv>=1.0.0
